tools/kbpy/frc/requirements.txt
pyfrc
robotpy-ctre
robotpy-navx
pynetworktables
numpy